Added swimming to 1st person, refactored state machine
Added swimming pool to 1st person character scene.  When in water, character moves slower and is unable to run, jump, or crouch.  Refactored 1st person character to eliminate reliance on MoveState, replaced with UnityEvent observer pattern.

+++ Assets/Playcraft/Tools/UnityEvent Helpers/Observer/Access/FloatEventAccess.cs
@@ -5,6 +5,6 @@ namespace Playcraft
     public class FloatEventAccess : MonoBehaviour
     {
         [SerializeField] FloatGameEvent invoker = default;
-        public void Input(int value) { invoker.Raise(value); }
+        public void Input(float value) { invoker.Raise(value); }
     }
 }
